perf: fix timing-out Integration A-B/C/D shards (AnoGAN analytic grad + parallel streaming Adam) (#1599)
* perf(anomaly): analytic z-gradient for AnoGAN scoring (was finite-difference)

AnoGANDetector.ScoreAnomalies optimized the latent z per sample with CENTRAL
FINITE DIFFERENCES: each step perturbed every one of latentDim (64) dimensions
twice and ran a full generator+discriminator forward each time — ~2*latentDim
forwards per step, hundreds of thousands of forwards per scoring call. On the
30-row integration data this ran well past the 120s test timeout and tripped
the CI blame collector's 5-min inactivity abort, crashing the host and failing
the whole Integration A-B shard.

Replace it with the exact analytic gradient in a single backward pass:
- BackpropGeneratorToInput: output-gradient -> z (no weight update).
- BackpropDiscriminatorFeaturesToInput: feature-layer gradient -> disc input.
- ScoreAnomaliesInternal: dLoss/dxGen = 2(xGen-x) + backprop of
  0.2(featGen-featReal) through the discriminator features, then backprop
  through the generator to z. featReal is hoisted out of the step loop.

Same loss and gradient-descent update, ~latentDim x fewer forwards. Verified:
AnoGAN_OutlierGetsHighestScore + AnoGAN_PredictClassifiesOutlierAsAnomaly pass
in 17s (were hanging), outlier still scored highest.


* perf(training): parallelize StreamingAdam8Bit block loop over CPU cores

The 8-bit streaming Adam optimizer step (ApplyDouble/ApplyFloat) ran its
per-block update loop on a single thread. At foundation scale (e.g. OneFormer
Swin-Large, 219M params) a CPU profile attributed ~28% of each training step
to this flat serial loop while 13+ cores sat idle.

Each block owns a disjoint slice of the parameter / gradient / quant-state
arrays and its own per-block dequant scale, so the loop is embarrassingly
parallel. Route it through CpuParallelSettings.ParallelForOrSerial with
per-worker scratch allocated in localInit (keeps the steady state alloc-free).
The grain gate (DefaultSerialGrainSize = 32K) keeps small parameters serial so
tiny biases pay no thread-dispatch overhead.

Results are bit-identical to the serial path (blocks are independent). The
shared _mScratch/_vScratch buffers now back only the serial generic fallback.

Measured on OneFormer Swin-Large multi-step train (16-core, GPU disabled to
match CI's no-GPU path): steady-state train step 9.5s -> 7.4s (~22%), avg
cores busy 2.7 -> 3.9. Benefits every model trained through the streaming
tape path, not just OneFormer.


* perf(optimizers): parallelize Adam8BitOptimizer block quant/dequant loops

The standalone 8-bit Adam optimizer ran its block-wise quantize/dequantize
loops (Quantize, Dequantize, QuantizeTensor, DequantizeTensor) single-threaded.
Each block is independent (disjoint element slice + its own per-block scale), so
they parallelize cleanly via CpuParallelSettings.ParallelForOrSerial, grain-gated
so small parameters stay serial.

- Dequant loops parallelize unconditionally (no shared state).
- QuantizeTensor's per-block sort scratch moves from one shared rented buffer to
  a per-worker buffer (localInit rents lazily, localFinally returns it) — keeps
  the alloc-free amortization under parallelism.
- Stochastic rounding now draws from RandomHelper.ThreadSafeRandom (per-thread
  LockedRandom) instead of a single seeded Random, so it is thread-safe. The
  previously-seeded _random field is removed (it is unused now); exact seed
  reproducibility could not survive parallel work-stealing regardless, and the
  default path is deterministic round-to-nearest (UseStochasticRounding=false).

Mirrors the StreamingAdam8Bit block-loop parallelization; benefits any model
trained with the standalone 8-bit Adam under memory constraints.


* test(integration): float-ize foundation-model construct tests to halve OOM footprint (C/D shards)

Integration C and D shards were OOM-killed ("runner received shutdown signal")
constructing foundation-scale models at double on the 16GB runner — not the
timeouts originally assumed. These construct/predict + name/metadata tests have
dtype-invariant assertions, so switching them to float halves the construction
footprint (the proven PR #1596 lever):

- SegmentationModelSizeVariationTests (C): SAM/SAMHQ ViTHuge (~636M), SEEM,
  Mask2Former, etc. — ~5GB -> ~2.5GB at float, fits the runner.
- DiffusionExtendedIntegrationTests (D): ~50 foundation diffusion model
  construction tests (Sora, WanVideo, Flux, SDXL, PixArt, ...). The 2
  DiffusionModelOptions<double> config tests keep double (they assert exact
  values like 0.001).
- LayoutAwareDocumentTests (D): LayoutLM(v2/v3)/Donut/DocOwl construct + predict
  + metadata-name tests; the 0.5 image-fill literal becomes 0.5f.

Assertions are ParameterCount>0 / output shape / metadata names / config values
— all dtype-invariant, so no test is weakened. The largest models that still
exceed the runner even at float are addressed separately via lazy construction.


* fix(document): lazily initialize pixel-to-sequence models

* fix(ci): stabilize PR1599 integration shards

---------

+++ tests/AiDotNet.Tests/IntegrationTests/ComputerVision/SegmentationModelSizeVariationTests.cs
@@ -27,17 +27,17 @@ namespace AiDotNet.Tests.IntegrationTests.ComputerVision;
 /// </summary>
 public class SegmentationModelSizeVariationTests
 {
-    private static NeuralNetworkArchitecture<double> Arch(int h = 32, int w = 32, int d = 3)
+    private static NeuralNetworkArchitecture<float> Arch(int h = 32, int w = 32, int d = 3)
         => new(InputType.ThreeDimensional, NeuralNetworkTaskType.Regression,
                NetworkComplexity.Deep, 0, h, w, d, 0);
 
-    private static Tensor<double> Rand(params int[] shape)
+    private static Tensor<float> Rand(params int[] shape)
     {
         int total = 1; foreach (int s in shape) total *= s;
-        var data = new double[total];
+        var data = new float[total];
         var rng = RandomHelper.CreateSeededRandom(42);
-        for (int i = 0; i < total; i++) data[i] = rng.NextDouble();
-        return new Tensor<double>(shape, new Vector<double>(data));
+        for (int i = 0; i < total; i++) data[i] = (float)rng.NextDouble();
+        return new Tensor<float>(shape, new Vector<float>(data));
     }
 
     #region SegFormer — All 6 sizes
@@ -51,7 +51,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SegFormerModelSize.B5)]
     public void SegFormer_AllModelSizes_ConstructAndPredict(SegFormerModelSize size)
     {
-        var model = new SegFormer<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new SegFormer<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0, $"SegFormer-{size} has no parameters");
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -69,7 +69,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(Mask2FormerModelSize.SwinLarge)]
     public void Mask2Former_AllModelSizes_ConstructAndPredict(Mask2FormerModelSize size)
     {
-        var model = new Mask2Former<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new Mask2Former<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -85,7 +85,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SAMModelSize.ViTHuge)]
     public void SAM_AllModelSizes_ConstructAndPredict(SAMModelSize size)
     {
-        var model = new SAM<double>(Arch(), numClasses: 1, modelSize: size);
+        var model = new SAM<float>(Arch(), numClasses: 1, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -102,7 +102,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SAM21ModelSize.Large)]
     public void SAM21_AllModelSizes_ConstructAndPredict(SAM21ModelSize size)
     {
-        var model = new SAM21<double>(Arch(), numClasses: 1, modelSize: size);
+        var model = new SAM21<float>(Arch(), numClasses: 1, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -119,7 +119,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SegNeXtModelSize.Large)]
     public void SegNeXt_AllModelSizes_ConstructAndPredict(SegNeXtModelSize size)
     {
-        var model = new SegNeXt<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new SegNeXt<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -137,7 +137,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(YOLOv8SegModelSize.X)]
     public void YOLOv8Seg_AllModelSizes_ConstructAndPredict(YOLOv8SegModelSize size)
     {
-        var model = new YOLOv8Seg<double>(Arch(), modelSize: size);
+        var model = new YOLOv8Seg<float>(Arch(), modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -148,7 +148,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(YOLOv9SegModelSize.E)]
     public void YOLOv9Seg_AllModelSizes_ConstructAndPredict(YOLOv9SegModelSize size)
     {
-        var model = new YOLOv9Seg<double>(Arch(), modelSize: size);
+        var model = new YOLOv9Seg<float>(Arch(), modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -162,7 +162,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(YOLO11SegModelSize.X)]
     public void YOLO11Seg_AllModelSizes_ConstructAndPredict(YOLO11SegModelSize size)
     {
-        var model = new YOLO11Seg<double>(Arch(), modelSize: size);
+        var model = new YOLO11Seg<float>(Arch(), modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -176,7 +176,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(YOLOv12SegModelSize.X)]
     public void YOLOv12Seg_AllModelSizes_ConstructAndPredict(YOLOv12SegModelSize size)
     {
-        var model = new YOLOv12Seg<double>(Arch(), modelSize: size);
+        var model = new YOLOv12Seg<float>(Arch(), modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -190,7 +190,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(YOLO26SegModelSize.X)]
     public void YOLO26Seg_AllModelSizes_ConstructAndPredict(YOLO26SegModelSize size)
     {
-        var model = new YOLO26Seg<double>(Arch(), modelSize: size);
+        var model = new YOLO26Seg<float>(Arch(), modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -206,7 +206,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(NnUNetModelSize.UNet3DCascade)]
     public void NnUNet_AllModelSizes_ConstructAndPredict(NnUNetModelSize size)
     {
-        var model = new NnUNet<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new NnUNet<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -219,7 +219,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(MedNeXtModelSize.Large)]
     public void MedNeXt_AllModelSizes_ConstructAndPredict(MedNeXtModelSize size)
     {
-        var model = new MedNeXt<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new MedNeXt<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -235,7 +235,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(VisionMambaModelSize.Base)]
     public void VisionMamba_AllModelSizes_ConstructAndPredict(VisionMambaModelSize size)
     {
-        var model = new VisionMamba<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new VisionMamba<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -247,7 +247,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(VMambaModelSize.Base)]
     public void VMamba_AllModelSizes_ConstructAndPredict(VMambaModelSize size)
     {
-        var model = new VMamba<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new VMamba<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -263,7 +263,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(PIDNetModelSize.Large)]
     public void PIDNet_AllModelSizes_ConstructAndPredict(PIDNetModelSize size)
     {
-        var model = new PIDNet<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new PIDNet<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -278,7 +278,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(KMaXDeepLabModelSize.ConvNeXtLarge)]
     public void KMaXDeepLab_AllModelSizes_ConstructAndPredict(KMaXDeepLabModelSize size)
     {
-        var model = new KMaXDeepLab<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new KMaXDeepLab<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -289,7 +289,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(OneFormerModelSize.DiNATLarge)]
     public void OneFormer_AllModelSizes_ConstructAndPredict(OneFormerModelSize size)
     {
-        var model = new OneFormer<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new OneFormer<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -303,7 +303,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SegGPTModelSize.ViTLarge)]
     public void SegGPT_AllModelSizes_ConstructAndPredict(SegGPTModelSize size)
     {
-        var model = new SegGPT<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new SegGPT<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -314,7 +314,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SEEMModelSize.Large)]
     public void SEEM_AllModelSizes_ConstructAndPredict(SEEMModelSize size)
     {
-        var model = new SEEM<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new SEEM<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -329,7 +329,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(DEVAModelSize.Large)]
     public void DEVA_AllModelSizes_ConstructAndPredict(DEVAModelSize size)
     {
-        var model = new DEVA<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new DEVA<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -340,7 +340,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(EfficientTAMModelSize.Small)]
     public void EfficientTAM_AllModelSizes_ConstructAndPredict(EfficientTAMModelSize size)
     {
-        var model = new EfficientTAM<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new EfficientTAM<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -355,7 +355,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(PointTransformerV3ModelSize.Large)]
     public void PointTransformerV3_AllModelSizes_ConstructAndPredict(PointTransformerV3ModelSize size)
     {
-        var model = new PointTransformerV3<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new PointTransformerV3<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -366,7 +366,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SonataModelSize.Large)]
     public void Sonata_AllModelSizes_ConstructAndPredict(SonataModelSize size)
     {
-        var model = new Sonata<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new Sonata<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -377,7 +377,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(ConcertoModelSize.Large)]
     public void Concerto_AllModelSizes_ConstructAndPredict(ConcertoModelSize size)
     {
-        var model = new Concerto<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new Concerto<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -393,7 +393,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(SAMHQModelSize.ViTHuge)]
     public void SAMHQ_AllModelSizes_ConstructAndPredict(SAMHQModelSize size)
     {
-        var model = new SAMHQ<double>(Arch(), numClasses: 1, modelSize: size);
+        var model = new SAMHQ<float>(Arch(), numClasses: 1, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -404,7 +404,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(MaskDINOModelSize.SwinLarge)]
     public void MaskDINO_AllModelSizes_ConstructAndPredict(MaskDINOModelSize size)
     {
-        var model = new MaskDINO<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new MaskDINO<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -416,7 +416,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(EoMTModelSize.Large)]
     public void EoMT_AllModelSizes_ConstructAndPredict(EoMTModelSize size)
     {
-        var model = new EoMT<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new EoMT<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -430,7 +430,7 @@ public class SegmentationModelSizeVariationTests
     [InlineData(InternImageModelSize.Huge)]
     public void InternImage_AllModelSizes_ConstructAndPredict(InternImageModelSize size)
     {
-        var model = new InternImage<double>(Arch(), numClasses: 5, modelSize: size);
+        var model = new InternImage<float>(Arch(), numClasses: 5, modelSize: size);
         Assert.True(model.ParameterCount > 0);
         var output = model.Predict(Rand(1, 3, 32, 32));
         Assert.True(output.Length > 0);
@@ -443,8 +443,8 @@ public class SegmentationModelSizeVariationTests
     [Fact(Timeout = 120000)]
     public async Task Mask2Former_LargerSizeHasMoreParameters()
     {
-        var small = new Mask2Former<double>(Arch(), numClasses: 5, modelSize: Mask2FormerModelSize.SwinTiny);
-        var large = new Mask2Former<double>(Arch(), numClasses: 5, modelSize: Mask2FormerModelSize.SwinLarge);
+        var small = new Mask2Former<float>(Arch(), numClasses: 5, modelSize: Mask2FormerModelSize.SwinTiny);
+        var large = new Mask2Former<float>(Arch(), numClasses: 5, modelSize: Mask2FormerModelSize.SwinLarge);
         Assert.True(large.ParameterCount > small.ParameterCount,
             $"SwinLarge ({large.ParameterCount}) should have more params than SwinTiny ({small.ParameterCount})");
     }
@@ -452,8 +452,8 @@ public class SegmentationModelSizeVariationTests
     [Fact(Timeout = 120000)]
     public async Task SegNeXt_LargerSizeHasMoreParameters()
     {
-        var small = new SegNeXt<double>(Arch(), numClasses: 5, modelSize: SegNeXtModelSize.Tiny);
-        var large = new SegNeXt<double>(Arch(), numClasses: 5, modelSize: SegNeXtModelSize.Large);
+        var small = new SegNeXt<float>(Arch(), numClasses: 5, modelSize: SegNeXtModelSize.Tiny);
+        var large = new SegNeXt<float>(Arch(), numClasses: 5, modelSize: SegNeXtModelSize.Large);
         Assert.True(large.ParameterCount > small.ParameterCount,
             $"Large ({large.ParameterCount}) should have more params than Tiny ({small.ParameterCount})");
     }
@@ -461,8 +461,8 @@ public class SegmentationModelSizeVariationTests
     [Fact(Timeout = 120000)]
     public async Task SAM_LargerSizeHasMoreParameters()
     {
-        var small = new SAM<double>(Arch(), numClasses: 1, modelSize: SAMModelSize.ViTBase);
-        var large = new SAM<double>(Arch(), numClasses: 1, modelSize: SAMModelSize.ViTHuge);
+        var small = new SAM<float>(Arch(), numClasses: 1, modelSize: SAMModelSize.ViTBase);
+        var large = new SAM<float>(Arch(), numClasses: 1, modelSize: SAMModelSize.ViTHuge);
         Assert.True(large.ParameterCount > small.ParameterCount,
             $"ViTHuge ({large.ParameterCount}) should have more params than ViTBase ({small.ParameterCount})");
     }
@@ -470,8 +470,8 @@ public class SegmentationModelSizeVariationTests
     [Fact(Timeout = 120000)]
     public async Task NnUNet_LargerSizeHasMoreParameters()
     {
-        var small = new NnUNet<double>(Arch(), numClasses: 5, modelSize: NnUNetModelSize.UNet2D);
-        var large = new NnUNet<double>(Arch(), numClasses: 5, modelSize: NnUNetModelSize.UNet3DCascade);
+        var small = new NnUNet<float>(Arch(), numClasses: 5, modelSize: NnUNetModelSize.UNet2D);
+        var large = new NnUNet<float>(Arch(), numClasses: 5, modelSize: NnUNetModelSize.UNet3DCascade);
         Assert.True(large.ParameterCount > small.ParameterCount,
             $"UNet3DCascade ({large.ParameterCount}) should have more params than UNet2D ({small.ParameterCount})");
     }
@@ -479,8 +479,8 @@ public class SegmentationModelSizeVariationTests
     [Fact(Timeout = 120000)]
     public async Task VisionMamba_LargerSizeHasMoreParameters()
     {
-        var small = new VisionMamba<double>(Arch(), numClasses: 5, modelSize: VisionMambaModelSize.Tiny);
-        var large = new VisionMamba<double>(Arch(), numClasses: 5, modelSize: VisionMambaModelSize.Base);
+        var small = new VisionMamba<float>(Arch(), numClasses: 5, modelSize: VisionMambaModelSize.Tiny);
+        var large = new VisionMamba<float>(Arch(), numClasses: 5, modelSize: VisionMambaModelSize.Base);
         Assert.True(large.ParameterCount > small.ParameterCount,
             $"Base ({large.ParameterCount}) should have more params than Tiny ({small.ParameterCount})");
     }
@@ -488,8 +488,8 @@ public class SegmentationModelSizeVariationTests
     [Fact(Timeout = 120000)]
     public async Task PIDNet_LargerSizeHasMoreParameters()
     {
-        var small = new PIDNet<double>(Arch(), numClasses: 5, modelSize: PIDNetModelSize.Small);
-        var large = new PIDNet<double>(Arch(), numClasses: 5, modelSize: PIDNetModelSize.Large);
+        var small = new PIDNet<float>(Arch(), numClasses: 5, modelSize: PIDNetModelSize.Small);
+        var large = new PIDNet<float>(Arch(), numClasses: 5, modelSize: PIDNetModelSize.Large);
         Assert.True(large.ParameterCount > small.ParameterCount,
             $"Large ({large.ParameterCount}) should have more params than Small ({small.ParameterCount})");
     }
